Move switcher_kis coordinator to separate module
