+++ homeassistant/components/switcher_kis/climate.py
@@ -35,8 +35,8 @@ from homeassistant.helpers.dispatcher import async_dispatcher_connect
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
-from . import SwitcherDataUpdateCoordinator
 from .const import SIGNAL_DEVICE_ADD
+from .coordinator import SwitcherDataUpdateCoordinator
 from .utils import get_breeze_remote_manager
 
 DEVICE_MODE_TO_HA = {
